Add a `FileField` to our `File` class to manipulate files the django way (#19201)

+++ src/olympia/reviewers/tests/test_utils.py
@@ -1,5 +1,4 @@
 from datetime import datetime, timedelta
-from unittest import mock
 from unittest.mock import patch
 
 from django.conf import settings
@@ -87,7 +86,7 @@ class TestReviewHelperBase(TestCase):
         assert scores[0].note_key == reviewed_type
 
     def remove_paths(self):
-        if not storage.exists(self.file.file_path):
+        if self.file.file and not storage.exists(self.file.file_path):
             storage.delete(self.file.file_path)
 
     def create_paths(self):
@@ -148,11 +147,16 @@ class TestReviewHelperBase(TestCase):
 
 # Those tests can call signing when making things public. We want to test that
 # it works correctly, so we set ENABLE_ADDON_SIGNING to True and mock the
-# actual signing call.
+# actual signing call below in setUp().
 @override_settings(ENABLE_ADDON_SIGNING=True)
-@mock.patch('olympia.lib.crypto.signing.call_signing', lambda f: None)
 class TestReviewHelper(TestReviewHelperBase):
     __test__ = True
+
+    def setUp(self):
+        super().setUp()
+        patcher = patch('olympia.reviewers.utils.sign_file')
+        self.addCleanup(patcher.stop)
+        self.sign_file_mock = patcher.start()
 
     def test_type_nominated(self):
         assert self.setup_type(amo.STATUS_NOMINATED) == 'extension_nominated'
@@ -1064,9 +1068,8 @@ class TestReviewHelper(TestReviewHelperBase):
         assert not flags.pending_rejection
         assert not flags.pending_rejection_by
 
-    @patch('olympia.reviewers.utils.sign_file')
-    def test_nomination_to_public(self, sign_mock):
-        sign_mock.reset()
+    def test_nomination_to_public(self):
+        self.sign_file_mock.reset()
         self.setup_data(amo.STATUS_NOMINATED)
         AutoApprovalSummary.objects.update_or_create(
             version=self.version, defaults={'verdict': amo.AUTO_APPROVED, 'weight': 101}
@@ -1086,16 +1089,15 @@ class TestReviewHelper(TestReviewHelperBase):
         approval_counter = AddonApprovalsCounter.objects.get(addon=self.addon)
         assert approval_counter.counter == 1
 
-        sign_mock.assert_called_with(self.file)
+        self.sign_file_mock.assert_called_with(self.file)
         assert storage.exists(self.file.file_path)
 
         assert self.check_log_count(amo.LOG.APPROVE_VERSION.id) == 1
 
         self._check_score(amo.REVIEWED_EXTENSION_MEDIUM_RISK)
 
-    @patch('olympia.reviewers.utils.sign_file')
-    def test_old_nomination_to_public_bonus_score(self, sign_mock):
-        sign_mock.reset()
+    def test_old_nomination_to_public_bonus_score(self):
+        self.sign_file_mock.reset()
         self.setup_data(amo.STATUS_NOMINATED, type=amo.ADDON_PLUGIN)
         self.version.update(nomination=self.days_ago(9))
 
@@ -1113,7 +1115,7 @@ class TestReviewHelper(TestReviewHelperBase):
         approval_counter = AddonApprovalsCounter.objects.get(addon=self.addon)
         assert approval_counter.counter == 1
 
-        sign_mock.assert_called_with(self.file)
+        self.sign_file_mock.assert_called_with(self.file)
         assert storage.exists(self.file.file_path)
 
         assert self.check_log_count(amo.LOG.APPROVE_VERSION.id) == 1
@@ -1122,9 +1124,8 @@ class TestReviewHelper(TestReviewHelperBase):
         # 2 days over the limit = 4 points
         self._check_score(amo.REVIEWED_ADDON_FULL, bonus=4)
 
-    @patch('olympia.reviewers.utils.sign_file')
-    def test_nomination_to_public_not_human(self, sign_mock):
-        sign_mock.reset()
+    def test_nomination_to_public_not_human(self):
+        self.sign_file_mock.reset()
         self.setup_data(amo.STATUS_NOMINATED, human_review=False)
 
         self.helper.handler.approve_latest_version()
@@ -1145,7 +1146,7 @@ class TestReviewHelper(TestReviewHelperBase):
         # human review field should be empty.
         assert approval_counter.last_human_review is None
 
-        sign_mock.assert_called_with(self.file)
+        self.sign_file_mock.assert_called_with(self.file)
         assert storage.exists(self.file.file_path)
 
         assert self.check_log_count(amo.LOG.APPROVE_VERSION.id) == 1
@@ -1153,15 +1154,17 @@ class TestReviewHelper(TestReviewHelperBase):
         # No request, no user, therefore no score.
         assert ReviewerScore.objects.count() == 0
 
-    @patch('olympia.reviewers.utils.sign_file')
-    def test_public_addon_with_version_awaiting_review_to_public(self, sign_mock):
-        sign_mock.reset()
+    def test_public_addon_with_version_awaiting_review_to_public(self):
+        self.sign_file_mock.reset()
         self.addon.current_version.update(created=self.days_ago(1))
         self.version = version_factory(
             addon=self.addon,
             channel=amo.RELEASE_CHANNEL_LISTED,
             version='3.0.42',
-            file_kw={'status': amo.STATUS_AWAITING_REVIEW},
+            file_kw={
+                'status': amo.STATUS_AWAITING_REVIEW,
+                'filename': 'webextension.xpi',
+            },
         )
         self.preamble = 'Mozilla Add-ons: Delicious Bookmarks 3.0.42'
         self.file = self.version.file
@@ -1199,7 +1202,7 @@ class TestReviewHelper(TestReviewHelperBase):
         assert approval_counter.counter == 2
         self.assertCloseToNow(approval_counter.last_human_review)
 
-        sign_mock.assert_called_with(self.file)
+        self.sign_file_mock.assert_called_with(self.file)
         assert storage.exists(self.file.file_path)
 
         assert self.check_log_count(amo.LOG.APPROVE_VERSION.id) == 1
@@ -1208,10 +1211,7 @@ class TestReviewHelper(TestReviewHelperBase):
         self.addon.reviewerflags.reload()
         assert not self.addon.reviewerflags.auto_approval_disabled_until_next_approval
 
-    @patch('olympia.reviewers.utils.sign_file')
-    def test_public_addon_with_version_need_human_review_to_public(
-        self, sign_file_mock
-    ):
+    def test_public_addon_with_version_need_human_review_to_public(self):
         self.old_version = self.addon.current_version
         self.old_version.update(created=self.days_ago(1), needs_human_review=True)
         self.version = version_factory(
@@ -1232,10 +1232,7 @@ class TestReviewHelper(TestReviewHelperBase):
         self.old_version.reload()
         assert not self.old_version.needs_human_review
 
-    @patch('olympia.reviewers.utils.sign_file')
-    def test_public_addon_with_auto_approval_temporarily_disabled_to_public(
-        self, sign_file_mock
-    ):
+    def test_public_addon_with_auto_approval_temporarily_disabled_to_public(self):
         AddonReviewerFlags.objects.create(
             addon=self.addon, auto_approval_disabled_until_next_approval=True
         )
@@ -1257,15 +1254,17 @@ class TestReviewHelper(TestReviewHelperBase):
         self.addon.reviewerflags.reload()
         assert not self.addon.reviewerflags.auto_approval_disabled_until_next_approval
 
-    @patch('olympia.reviewers.utils.sign_file')
-    def test_public_addon_with_version_awaiting_review_to_sandbox(self, sign_mock):
-        sign_mock.reset()
+    def test_public_addon_with_version_awaiting_review_to_sandbox(self):
+        self.sign_file_mock.reset()
         self.addon.current_version.update(created=self.days_ago(1))
         self.version = version_factory(
             addon=self.addon,
             channel=amo.RELEASE_CHANNEL_LISTED,
             version='3.0.42',
-            file_kw={'status': amo.STATUS_AWAITING_REVIEW},
+            file_kw={
+                'status': amo.STATUS_AWAITING_REVIEW,
+                'filename': 'webextension.xpi',
+            },
         )
         self.preamble = 'Mozilla Add-ons: Delicious Bookmarks 3.0.42'
         self.file = self.version.file
@@ -1273,7 +1272,6 @@ class TestReviewHelper(TestReviewHelperBase):
         AutoApprovalSummary.objects.create(
             version=self.version, verdict=amo.AUTO_APPROVED, weight=101
         )
-        self.create_paths()
         AddonApprovalsCounter.objects.create(addon=self.addon, counter=1)
 
         # Safeguards.
@@ -1298,7 +1296,7 @@ class TestReviewHelper(TestReviewHelperBase):
         approval_counter = AddonApprovalsCounter.objects.get(addon=self.addon)
         assert approval_counter.counter == 1
 
-        assert not sign_mock.called
+        assert not self.sign_file_mock.called
         assert storage.exists(self.file.file_path)
         assert self.check_log_count(amo.LOG.REJECT_VERSION.id) == 1
 
@@ -1604,9 +1602,8 @@ class TestReviewHelper(TestReviewHelperBase):
         activity = activities[1]
         assert activity.arguments == [self.addon, second_unlisted]
 
-    @patch('olympia.reviewers.utils.sign_file')
-    def test_null_to_public_unlisted(self, sign_mock):
-        sign_mock.reset()
+    def test_null_to_public_unlisted(self):
+        self.sign_file_mock.reset()
         self.setup_data(amo.STATUS_NULL, channel=amo.RELEASE_CHANNEL_UNLISTED)
 
         self.helper.handler.approve_latest_version()
@@ -1627,15 +1624,14 @@ class TestReviewHelper(TestReviewHelperBase):
         )
         assert 'You received this email because' not in message.body
 
-        sign_mock.assert_called_with(self.file)
+        self.sign_file_mock.assert_called_with(self.file)
         assert storage.exists(self.file.file_path)
 
         assert self.check_log_count(amo.LOG.APPROVE_VERSION.id) == 1
 
-    @patch('olympia.reviewers.utils.sign_file')
-    def test_nomination_to_public_failed_signing(self, sign_mock):
-        sign_mock.side_effect = Exception
-        sign_mock.reset()
+    def test_nomination_to_public_failed_signing(self):
+        self.sign_file_mock.side_effect = Exception
+        self.sign_file_mock.reset()
         self.setup_data(amo.STATUS_NOMINATED)
 
         with self.assertRaises(Exception):
@@ -1651,8 +1647,7 @@ class TestReviewHelper(TestReviewHelperBase):
         assert len(mail.outbox) == 0
         assert self.check_log_count(amo.LOG.APPROVE_VERSION.id) == 0
 
-    @patch('olympia.reviewers.utils.sign_file')
-    def test_nomination_to_sandbox(self, sign_mock):
+    def test_nomination_to_sandbox(self):
         self.setup_data(amo.STATUS_NOMINATED)
         self.helper.handler.reject_latest_version()
 
@@ -1667,7 +1662,7 @@ class TestReviewHelper(TestReviewHelperBase):
         # AddonApprovalsCounter was not touched since we didn't approve.
         assert not AddonApprovalsCounter.objects.filter(addon=self.addon).exists()
 
-        assert not sign_mock.called
+        assert not self.sign_file_mock.called
         assert storage.exists(self.file.file_path)
         assert self.check_log_count(amo.LOG.REJECT_VERSION.id) == 1
